% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meteo_3DFEM.R
\name{check_meteo_3DFEM}
\alias{check_meteo_3DFEM}
\title{Check meteo data input as required by the 3DCMCCFEM model}
\usage{
check_meteo_3DFEM(
  file_meteo,
  make_report = FALSE,
  correct = TRUE,
  save = FALSE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{file_meteo}{Character or data.frame: path to meteo.txt file or
data.frame of meteo.txt file read with read.table.}

\item{make_report}{Logical: make a report of the input meteo file? To be
implemented yet.}

\item{correct}{Logical: correct any errors found? Default to TRUE.}

\item{save}{Logical: save meteo file?. Only if file_meteo is a path to a
file. If TRUE, meteo file is saved with the same name and same directory of
file_meteo. Default to FALSE}

\item{plot}{Logical: plot meteo file with \code{plot_meteo_3DFEM}?. Default
to FALSE.}

\item{...}{additional argument passed to \code{plot_meteo_3DFEM} when
\code{plot=TRUE}}
}
\value{
a data.frame with the corrected weather time series if
\code{correct=TRUE}, invisible otherwise. If \code{plot=TRUE} the function
return a ggplot object if \code{save_plot=FALSE}. If \code{save_plot=TRUE}
the function save in \code{outdir} a png file with the name of the site in
\code{meteo_path_file}
}
\description{
\code{check_meteo_3DFEM} perform some basic test on the variables in the
input meteo file (e.g., generated by the function \code{make_meteo_EOBS}) to
ensure compatibility with \code{run_3DFEM}. The function can also try to
correct inconsistency
}
\details{
The function check for consistency with the columns name convention
of the model, for NA, missing data, missing observation in the weather time
series, anomalous values and temperature relationships (i.e.,
Tmin<=Tmean<=Tmax). If \code{correct=TRUE} colnames and temperature are
fixed according to model specifications. After all checks, if
\code{plot=TRUE} the function \code{plot_meteo_3DFEM} is call.
}
\seealso{
\code{\link[=run_3DFEM]{make_meteo_EOBS}} and
\code{\link[=run_3DFEM]{plot_meteo_3DFEM}}
}
