% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gif_3DFEM.R
\name{make_gif_3DFEM}
\alias{make_gif_3DFEM}
\title{Create a gif from the output of 3DCMCCFEM model}
\usage{
make_gif_3DFEM(
  path,
  years = NULL,
  distribution = c("regular", "pseudorandom", "random"),
  outdir,
  return_gif = TRUE,
  delete_frame = TRUE
)
}
\arguments{
\item{path}{Character: path to simulation output file. Must be annual.}

\item{years}{Numeric vector: years of simulations from which to create the
gif}

\item{distribution}{Character: How tree should be distributed in the stand?.
Options are c("regular","pseudorandom","random"). The pseudorandom option
is the slowest but ensure a minimum distance among trees in the random distribution}

\item{outdir}{Character: path to the output directory.}

\item{return_gif}{Logical: Should a file gif be created? Default to TRUE.}

\item{site}{Character: site name for constructing the output name.}

\item{delate_frame}{Logical: Should the single images frames be deleted after
the creation of the gif file? default to FALSE}
}
\value{
Used for its side effects.It save in \code{outdir} as many images as
the length of the simulation (if \code{delete_frame=FALSE}) and a gif file (if
\code{return_gif=TRUE})
}
\description{
\code{make_gif_3DFEM} use the output file generated by the function
\code{run_3DFEM} to create a gif of the simulation. In particular each frame
will correspond to one simulation year. For this reason this function works only with the annual output of the model.
}
\details{
The function relys on \code{rgl::snapshot(...,top=TRUE)} so during
the run it is a good idea to not alter the screen, to avoid
\code{rgl::snapshot} capturing a wrong screenshot. The function became very
slow when the number of tree to plot exeeds 300.
}
\seealso{
\code{\link[=run_3DFEM]{run_3DFEM}}
}
