% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_meteo_EOBS.R
\name{make_meteo_EOBS}
\alias{make_meteo_EOBS}
\title{Create meteo input file for 3DCMCCFEM model}
\usage{
make_meteo_EOBS(
  plot,
  crs = NULL,
  buffer = 100,
  id = NULL,
  daterange = c("1980-01-01", "2022-06-30"),
  convert_rad = TRUE,
  parallel = FALSE,
  site = NULL,
  outdir = NULL,
  return_df = FALSE
)
}
\arguments{
\item{plot}{Sf object or something that can be converted to an sf point object
(i.e., a data.frame of coordinates). If plot is a data.frame the argument
crs must be specified.}

\item{crs}{Numeric: a valid EPSG code}

\item{buffer}{Numeric: buffer around plot to extract meteo data from}

\item{id}{Character: name of the column in \code{plot} that identifies different
points}

\item{daterange}{Character vector of length 2: starting and ending date from
which to extract climate data. Must be in the form yyyy-mm-dd}

\item{convert_rad}{Logical: Should the solar radiation be converted from
radiant to Wh/m-2? Default to TRUE.}

\item{parallel}{Logical: Should the extraction be performed in parallel? Default to FALSE}

\item{site}{Character: the name of the plot site. Used to construct the output name}

\item{outdir}{Character: path to the output directory.}

\item{return_df}{Logical: should the climate data be returned by the function
as a data.frame? Default to FALSE}
}
\value{
Used for its side effects.It save in \code{outdir} a file
site_meteo.txt The output file have the colums:
\itemize{
\item \emph{Year}: Reference year for the meteo data
\item \emph{Month}: number of month in the year
\item \emph{n_day}: number of day in the month
\item \emph{Rg_f}: Solar radiation on the ground (in  Wh/m-2)
\item \emph{Ta_f}: Mean temperature (in 째C)
\item \emph{Tmax}: Maximum temperature (in 째C)
\item \emph{Tmin}: Minimum temperature (in 째C)
\item \emph{RH_f}: relative humidity (in \%)
\item \emph{Ts_f}: Daily soil temperature (in 째C)
\item \emph{Precip}: Daily precipitation (in mm day-1)
\item \emph{SWC}:Soil Water Content (in mm m-2)*
\item \emph{LAI}:Leaf Area Index (in m2 m-2) *
\item \emph{ET}: Evapotranspiration (in mm m-2 day-1*
\item \emph{WS_f}: Wind speed (in m s-1)*
}
}
\description{
\code{make_meteo_EOBS} use the
\href{https://surfobs.climate.copernicus.eu/dataaccess/access_eobs.php#datafiles/}{E-OBS
database} to create the meteo input file according to the 3DCMCCFEM model
specifications. The time period covered by E-OBS is 1950-01-01 - 2022-06-30.
}
\details{
This function use the version 26.0 of the E-OBS climate data to
generate the input file for the 3D-CMCC-FEM model. If parallel=TRUE, user
ca register his own backhand, otherwise the function will register a
7-workers implicit cluster based on \code{\link[foreach]{foreach}} and
\code{\link[doParallel]{registerDoParallel}}. For more information on meteo
file needed by the model please refer to the \href{
http://eprints.bice.rm.cnr.it/22393/1/3D-CMCC-FEM_User_Guide_v.2_July_2023.pdf}{User's
Guide}
}
\seealso{
\code{\link{run_3DFEM}}
}
