% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_soil_ISRIC.R
\name{make_soil_ISRIC}
\alias{make_soil_ISRIC}
\title{Create soil input file for 3DCMCCFEM model}
\usage{
make_soil_ISRIC(
  plot,
  crs = NULL,
  buffer = 100,
  id = NULL,
  site = NULL,
  outdir = NULL,
  return_df = FALSE
)
}
\arguments{
\item{plot}{Sf object or something that can be converted to an sf point object
(i.e., a data.frame of coordinates). If plot is a data.frame the argument
crs must be specified.}

\item{crs}{Numeric: a valid EPSG code}

\item{buffer}{Numeric: buffer around plot to extract meteo data from}

\item{id}{Character: name of the column in \code{plot} that identifies different
points}

\item{site}{Character: the name of the plot site. Used to construct the output name}

\item{outdir}{Character: path to the output directory.}

\item{return_df}{Logical: should the climate data be returned by the function
as a data.frame? Default to FALSE}
}
\value{
Used for its side effects.It save in \code{outdir} a file
site_soil.txt. The most important columns in this file are:
\itemize{
\item \emph{X}: x cell position
\item \emph{Y}: y cell position
\item \emph{LANDUSE}: "F" by default: Forest land use.
\item \emph{LAT}: latitude of point
\item \emph{LON}: longitude of point
\item \emph{CLAY_PERC}: soil clay (in \%)
\item \emph{SILT_PERC}: soil silt (in \%)
\item \emph{SAND_PERC}: soil sand (in \%)
\item \emph{SAND_PERC}: soil sand (in \%)
\item \emph{SOIL_DEPTH}: soil depth (in m)
\item \emph{FR}:0.9 by default. Fertility rating (dim) (only LUE version)
\item \emph{FN0}:0.5 by default. Value of fertility modifier when FR=0 (dim) (only LUE version)
\item \emph{M0}:0.2 by default. Value of ‘m’ when FR=0 (dim) (only LUE version)
}
}
\description{
\code{make_soil_ISRIC} use the
\href{https://www.isric.org/explore/soilgrids}{soilGrids} from the ISRIC
database to create the soil input file according to the 3DCMCCFEM model
specifications.
}
\details{
This function compute the weighted mean percentage of sand, silt and
clay in the first meter of soil. For more information on soil file needed
by the model please refer to the \href{
http://eprints.bice.rm.cnr.it/22393/1/3D-CMCC-FEM_User_Guide_v.2_July_2023.pdf}{User's
Guide}
}
\seealso{
\code{\link{run_3DFEM}}
}
