% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stand_INFC.R
\name{make_stand_INFC}
\alias{make_stand_INFC}
\title{Create stand input file for 3DCMCCFEM model}
\usage{
make_stand_INFC(
  id,
  mode = c("tree-level", "stand-level"),
  year = c("2005", "2015", "both"),
  make_topo = FALSE,
  site = NULL,
  outdir = NULL
)
}
\arguments{
\item{id}{Numeric vector: IDs of INFC plot.}

\item{mode}{Character: one of "tree-level", "stand-level". From which
database should the data be taken?}

\item{year}{Character: one of "2005","2015","both". From which year should
the information be taken?}

\item{make_topo}{Logical: If TRUE use the function \code{make_topo} to create
also the topo file according to the INFC database}

\item{site}{Character: Name of the site. Used to construct output name}

\item{outdir}{Character: path to the output directory.}
}
\value{
Used for its side effects.It save in \code{outdir} as many stand
input files as the length of the IDs provided
}
\description{
\code{make_stand_INFC} use Italian INFC 2005 and 2015 data to produce the
stand.txt file needed by the function 3DFEM. The function can use both
stand-level and tree-level data. More information about INFC data can be
found  \href{https://www.inventarioforestale.org/it/}{here}
}
\details{
The function can accept a vector of IDs and return as many files as
the length of the IDs vector.The output stand file has the columns:
\itemize{
\item \emph{Year}: Reference year for the data
\item \emph{x}: x cell position
\item \emph{y}: y cell position
\item \emph{Age}: Age of the stand
\item \emph{Species}: Species present within the stand
\item \emph{Managment}: Is the stand managed: Logical
\item \emph{N}: Numbers of trees per hectare
\item \emph{Stools}: Number of stool per cell*
\item \emph{AvDBH}: Average Diameter at Breast Height for the species (in
cm)
\item \emph{Height}: Avarage height of the species (in m)
\item \emph{Wf}:Foliage biomass*
\item \emph{Wrc}: Coarse root biomass *
\item \emph{Wrf}: Fine root biomass*
\item \emph{Ws}: Steam biomass*
\item \emph{Wbb}: Branch and bark biomass*
\item \emph{Wres}: Reserve biomass*
\item \emph{Lai}: Leaf Area Index of the stand*
}
All the data in the input file are referred to the dimension of the INFC
plot (i.e., 530 m2). Input signed with * are not mandatory for the model.
The stand file generated by the function will have a row for each year (if
year="both") and each species find in the plot (when mode="tree-level) or a
single row with average plot data (when mode="stand-level"). The age is
always retrieve from the stand-level database. For more info on the stand
input data needed see the official \href{
http://eprints.bice.rm.cnr.it/22393/1/3D-CMCC-FEM_User_Guide_v.2_July_2023.pdf}{User's
Guide}
}
\seealso{
\code{\link[=run_3DFEM]{run_3DFEM}}
}
