% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meteo_3DFEM.R
\name{plot_meteo_3DFEM}
\alias{plot_meteo_3DFEM}
\title{Plot meteo data input as required by the 3DCMCCFEM model}
\usage{
plot_meteo_3DFEM(
  meteo_file_path = NULL,
  var = "all",
  daterange = c("1980-01-01", "2022-06-30"),
  window = NULL,
  save_plot = FALSE,
  outdir = NULL
)
}
\arguments{
\item{meteo_file_path}{Character: path to simulation output file.}

\item{var}{Character vector: vector of variables name to be plotted. Default
to "all" to plot all variables.}

\item{daterange}{Character vector of length 2: starting and ending date from
which to plot meteo data. Must be in the form yyyy-mm-dd. Default to the
time range of the EOBS dataset}

\item{window}{Numeric: dimension of the focal window for compute the rolling
mean. Expressed in days. Deafult to NULL (rolling mean is not plotted).}

\item{save_plot}{Logical: Should plots be saved in outdir? Deafult to FALSE.}

\item{outdir}{Character: path to the output directory.}
}
\value{
A ggplot object if \code{save_plot=FALSE}, invisible otherwise. If
\code{save_plot=TRUE} the function save in \code{outdir} a png file with
the name of the site in \code{meteo_path_file}
}
\description{
\code{plot_meteo_3DFEM} plot the variables in the input meteo file (e.g.,
generated by the function \code{make_meteo_EOBS}) required by the model.
}
\details{
The function plots the time series of the specified variables, as
reported from  \code{make_meteo_EOBS}. It tries to guess the best time
scale for the x-axis based on the \code{daterange} argument. The
\code{window} argument should be appropriate to the desired time frame.
}
\seealso{
\code{\link[=run_3DFEM]{make_meteo_EOBS}}
}
