% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output_3DFEM.R
\name{plot_output_3DFEM}
\alias{plot_output_3DFEM}
\title{Plot the annual output of the 3DCMCCFEM model}
\usage{
plot_output_3DFEM(
  output_3DFEM,
  x = NULL,
  y = NULL,
  site_name = NULL,
  year_range = NULL,
  window = NULL,
  text_size = 40,
  save_plot = FALSE,
  outdir = NULL
)
}
\arguments{
\item{output_3DFEM}{Character: path to the model output file.}

\item{x}{Character vector of length 1: variable name to be plotted in the
x-axis. If not provided will be set to the variable "YEAR"}

\item{y}{Character vector: variable name to be plotted in the y-axis. Could
be of length >1 if x is NULL. In this case variables will be plotted on
different facets}

\item{site_name}{character vector of length 1: site name to be plotted as
title. Optional}

\item{year_range}{numeric vector of length 2: starting and ending year from
which to plot the data. It will be sorted in ascending order}

\item{window}{Numeric: dimension of the focal window for compute the rolling
mean. Expressed in years Default to NULL (rolling mean is not plotted).}

\item{save_plot}{Logical: Should plots be saved in outdir? Default to FALSE.}

\item{outdir}{Character: path to the output directory.}
}
\value{
A ggplot object if \code{save_plot=FALSE}, invisible otherwise. If
\code{save_plot=TRUE} the function save in \code{outdir} a png file. The
file name will be constructed based on the argument provided (x,
y, site_name, year_range, and window)
}
\description{
\code{plot_output_3DFEM} plot the variables in the output file of the model
(e.g., generated by the function \code{run_3DFEM}).
}
\details{
The function plots the time series of the variables specified by y.
The \code{window} argument should be appropriate to the desired time frame.
If x is provided, a scatterplot of the x vs y variable will be plotted and
the window argument will be ignored.
}
\seealso{
\code{\link[=run_3DFEM]{run_3DFEM}}
}
