% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_soil_3DFEM.R
\name{plot_soil_3DFEM}
\alias{plot_soil_3DFEM}
\title{Plot soil data input as required by the 3DCMCCFEM model}
\usage{
plot_soil_3DFEM(
  soil_file_path = NULL,
  save_plot = TRUE,
  outdir = NULL,
  main = "",
  show.names = TRUE,
  show.lines = TRUE,
  col.names = "gray",
  bg.names = par("bg"),
  show.grid = FALSE,
  col.axis = "black",
  col.lines = "gray",
  col.grid = "gray",
  lty.grid = 3,
  show.legend = FALSE,
  label.points = FALSE,
  point.labels = NULL,
  col.symbols = "red",
  pch = par("pch"),
  ...
)
}
\arguments{
\item{soil_file_path}{Character vector: vector of paths to soil.txt input files.}

\item{save_plot}{Logical: Should plots be saved in outdir? Deafult to FALSE.}

\item{outdir}{Character: path to the output directory.}

\item{main}{The title of the triangle plot. Defaults to nothing.}

\item{show.grid}{Whether to display grid lines at the ticks.}

\item{col.axis}{Color of the triangular axes, ticks and labels.}

\item{col.grid}{Color of the grid lines. Defaults to gray.}

\item{lty.grid}{Type of line for the grid.}

\item{show.legend}{Logical - whether to display a legend.}

\item{label.points}{Logical - whether to call \samp{thigmophobe.labels} to
  label the points.}

\item{point.labels}{Optional labels for the points and/or legend.}

\item{col.symbols}{Color of the symbols representing each value.}

\item{pch}{Symbols to use in plotting values.}

\item{...}{Additional arguments passed to \samp{points}.}
}
\value{
Invisible. The function is called for its side effects. If
\code{save_plot=TRUE} the function save in \code{outdir} a png file with
the name of the site in \code{soil_file_path}
}
\description{
\code{plot_soil_3DFEM} plot the textural information of a soil.txt file (e.g.,
generated by the function \code{make_soil_ISRIC}) required by the model.
Textural information is represented as the textural triangle with the USDA
system.
}
\details{
The function plots the textural characteristics of soil from the
input soil.txt file required by the 3D-CMCC-FEM model. Based on the
\code{plotrix} package. See plotrix::triax.plot for all possible arguments.
}
\seealso{
\code{\link[=plot_meteo_3DFEM]{plot_meteo_3DFEM}}
}
