% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R3DFEM.R
\docType{package}
\name{R3DFEM}
\alias{R3DFEM}
\title{R3DFEM: An package for running the 3D-CMCC-FEM model.}
\description{
The R3DFEM package provides functions for creating and check inputs for
simulations, run the 3D-CMCC-FEM model, plot inputs and outputs and other
functionalities. See the package readme for an overview of the package
functionality.
}
\note{
See the
\href{https://www.forest-modelling-lab.com/the-3d-cmcc-model}{official web page
} for a complete documentation of the model.
}
\author{
Elia Vangi
}
