% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_3DFEM.R
\name{run_3DCMCCFEM}
\alias{run_3DCMCCFEM}
\title{Wrapper for the 3D-CMCC-FEM model}
\usage{
run_3DCMCCFEM(
  site = NULL,
  species = NULL,
  year_start = NULL,
  year_end = NULL,
  co2 = "on",
  man = NULL,
  pres_dens = NULL,
  after_pres = NULL,
  reg = "off",
  year_start_man = -9999,
  rep_trees = -9999,
  rep_age = -9999,
  rep_dbh = -9999,
  rep_lai = -9999,
  rep_height = -9999,
  rep_species = -9999,
  reg_trees = -9999,
  reg_age = -9999,
  reg_dbh = -9999,
  reg_lai = -9999,
  reg_height = -9999,
  output = NULL,
  ps_accl = "on",
  rs_accl = "on",
  psn = "BGC",
  co2_conc = 368.865,
  frac_maxasw = 1,
  cell = 100,
  pruning = "off",
  irrigation = "off",
  inputdir = NULL,
  outputdir = NULL
)
}
\arguments{
\item{site}{Character: Site name. Used to construct output name}

\item{species}{Character: Species names (same as parameterization files)}

\item{year_start}{Integer: Starting year of simulation}

\item{year_end}{Integer: Ending year of simulation}

\item{co2}{Character: CO2 concentration changes according to scenario;
options: c("on","off","var")}

\item{man}{Character: Management options; options: c("on","off","var")}

\item{pres_dens}{Character: Prescribed densities in stand file if man =
"var"; options: c("on","off")}

\item{after_pres}{Character: Management after last year of prescribed density
if man = "var"; options: c("on","off")}

\item{reg}{Character: Regeneration; options: c("on","off")}

\item{year_start_man}{Integer: If man = "on"; Starting year of management;
Default to -9999.}

\item{rep_trees}{Integer: Number of replanted trees; Default to -9999.}

\item{rep_age}{Integer: Age of replanted trees; Default to -9999.}

\item{rep_dbh}{Integer: Average DBH of replanted trees; Default to -9999.}

\item{rep_lai}{Integer: Average LAI of replanted trees; Default to -9999.}

\item{rep_height}{Integer: Average height of replanted trees; Default to
-9999.}

\item{rep_species}{Integer: Species of replanted trees}

\item{reg_trees}{Integer: Number of saplings; Default to -9999.}

\item{reg_age}{Integer: Age of saplings; Default to -9999.}

\item{reg_dbh}{Integer: Average DBH of saplings; Default to -9999.}

\item{reg_lai}{Integer: Average LAI of saplings; Default to -9999.}

\item{reg_height}{Integer: Average height of saplings; Default to -9999.}

\item{output}{Character: Output type; options: c("daily","monthly","annual");}

\item{ps_accl}{Character: Photosynthesis acclimation; options : c("on","off")}

\item{rs_accl}{Character: Respiration acclimation; options : c("on","off");}

\item{psn}{Character: Photosynthesis version (Farquhar, von Caemmerer & Berry
or Light Use Efficiency); options: c("BGC","LUE")}

\item{co2_conc}{Numeric: Concentration of CO2 if co2  = "off"; Default to
368.865}

\item{frac_maxasw}{Numeric: Percentage of water content in soil compared to
maximum at the beginning of simulation (1 = 100\%); Default to 1}

\item{cell}{Numeric: Size of the cell; Default to 100 (1 ha)}

\item{pruning}{Character: Do pruning be performed? options: c("on","off");
Default to "off}

\item{irrigation}{Character: Do irrigation be performed? options:
c("on","off"); Default to "off}

\item{inputdir}{Character: Path to input files directory}

\item{outputdir}{Character: Path to output files directory}

\item{#}{Set regeneration parameters}
}
\value{
Used for its side effects.It save in \code{outdir} a .txt file with
the results of the simulations performed. For all details of the output
please refer to the  \href{
http://eprints.bice.rm.cnr.it/22393/1/3D-CMCC-FEM_User_Guide_v.2_July_2023.pdf}{User's
Guide}
}
\description{
\code{run_3DCMCCFEM} is the wrapper for the 3D-CMCC-FEM model developed by the
Italian National Research Council (CNR). Originally written in C, the current
version implemented in this package is the 5.6 ISIMIP. For all details on
the model please refer to the
\href{https://www.forest-modelling-lab.com/the-3d-cmcc-model}{official web
page}. For the development version and original code please refer to the
\href{https://github.com/Forest-Modelling-Lab/3D-CMCC-FEM}{GitHub repository}
}
\details{
The function need at least 6 input .txt files, that must be in the same
directory and should be named according to the name of the site or the
species: "site_stand.txt", "site_soil.txt", "site_topo.txt",
"site_meteo.txt", "site_co2.txt", "species.txt". A "site_management.txt"
file must be provided if required. The site name and species in the input
file name must be the same of the argument \code{site} and \code{species}.
}
\seealso{
\code{\link{virtual_stand_generator}}
}
